feat(database): enhance query execution with context support and improve type handling
- Add context support to ExecuteQueryStatement via new Option pattern
- Implement proper zero value returns for all column types (blob, bool, date, double, float, int32, int64)
- Replace stdBatch with stdStmt that implements full driver.Stmt interface
- Fix parameter binding regex to properly handle escaped question marks
- Improve connection pooling with timezone configuration and proper session management
- Add transaction support reporting with appropriate error messages
- Implement proper NULL value handling in row scanning
- Add comprehensive test coverage for new features and edge cases
- Introduce Options pattern for flexible client configuration
- Enhance session lifecycle management with proper resource cleanup

+++ database/column/double.go
@@ -38,14 +38,14 @@ func (d *Double) Row(stat *client.SessionDataSet, ptr bool) any {
 		if ptr {
 			return nil
 		}
-		return 0
+		return float64(0)
 	}
 	value, err := stat.GetDouble(d.name)
 	if err != nil {
 		if ptr {
 			return nil
 		}
-		return 0
+		return float64(0)
 	}
 	if ptr {
 		return &value
